style: 테스트 코드의 클래스 변경
