+++ src/test/java/racingcar/model/RacerTest.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import racingcar.constant.message.Message;
 import racingcar.dto.Result;
@@ -16,26 +15,8 @@ import java.util.List;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class RacerTest {
-
-    @DisplayName("잘못된 참가자 입력 체크")
-    @ParameterizedTest(name = "{displayName}: {0}")
-    @NullSource
-    @ValueSource(strings = {"", ",", ",car", ",,,,,", ",,dk", "a,b,c,,", "ㄱ,ㄴ,,,"})
-    void checkRacer(String value) {
-        assertThatThrownBy(() -> Racer.of(value))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("중복된 이름 체크")
-    @ParameterizedTest(name = "{displayName}: {0}")
-    @ValueSource(strings = {"ad,k,la,a,la", "l,l", "qwe,kz,pi,pi", "a, a,a ", "ba,l, ba"})
-    void checkUnique(String value) {
-        assertThatThrownBy(() -> Racer.of(value))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
 
     @DisplayName("올바른 이름 입력 체크")
     @ParameterizedTest(name = "{displayName}: {0}, {1}")
